+++ amazon/dynamodb/client/src/main/kotlin/org/http4k/connect/amazon/dynamodb/model/extensions.kt
@@ -1,6 +1,7 @@
 package org.http4k.connect.amazon.dynamodb.model
 
 import org.http4k.connect.amazon.dynamodb.model.DynamoDataType.valueOf
+import org.http4k.lens.BiDiMapping
 
 typealias TokensToNames = Map<String, AttributeName>
 typealias TokensToValues = Map<String, AttributeValue>
@@ -22,4 +23,25 @@ fun <T> Attribute<T>.asKeySchema(keyType: KeyType) = KeySchema(name, keyType)
  * Used for creating tables
  */
 fun <T> Attribute<T>.asAttributeDefinition() = AttributeDefinition(name, dataType)
+
+/**
+ * Map items out of a collection
+ */
+fun <T> Attribute.AttrLensSpec<List<AttributeValue>>.map(next: BiDiMapping<Item, T>): Attribute.AttrLensSpec<List<T>> = map(
+    { it.mapNotNull { it.M }.map { next(it) } },
+    { it.map { AttributeValue.Map(next(it)) } },
+)
+
+/**
+ * Map items out of a collection
+ */
+fun <T> Attribute.Companion.list(next: BiDiMapping<Item, T>) = list().map(
+    { it.mapNotNull { it.M }.map { next(it) } },
+    { it.map { AttributeValue.Map(next(it)) } },
+)
+
+/**
+ * Map items out of a collection
+ */
+fun <T> Attribute.Companion.map(next: BiDiMapping<Item, T>) = map().map(next)