+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 This list is not intended to be all-encompassing - it will document major and breaking API 
 changes with their rationale when appropriate. Given version `A.B.C.D`, breaking changes are to be expected in version number increments where changes in the `A` or `B` sections:
 
+### v3.22.2.0
+- **http4k-connect-*** - Upgrade dependencies.
+- **http4k-connect-amazon-dynamo*** - Added some useful methods for mapping Attributes such as `Attribute.map(BiDiMapping)` and `Attribute.list(BiDiMapping)`
+
 ### v3.22.1.0
 - **http4k-connect-*** - Upgrade dependencies.